fix portals dying
due to nummasks being too small and causing an overflow
ty hannu!

+++ src/r_things.h
@@ -81,7 +81,7 @@ typedef struct
 	sector_t* viewsector;
 } maskcount_t;
 
-void R_DrawMasked(maskcount_t* masks, UINT8 nummasks);
+void R_DrawMasked(maskcount_t* masks, INT32 nummasks);
 
 boolean R_ThingVisible(mobj_t *thing);
 boolean R_ThingWithinDist(mobj_t *thing, INT32 limit_dist);
